packages/memory/pyproject.toml
[project]
name = "sovereign-edge-memory"
version = "0.3.1"
requires-python = ">=3.11"
dependencies = [
    "sovereign-edge-core",
    "httpx>=0.27,<1.0",
    "lancedb>=0.30,<1.0",
    "numpy>=1.26,<3.0",
    "ollama>=0.4,<1.0",
]

[project.optional-dependencies]
# mem0ai==0.1.13 caps openai<2.0.0 but litellm==1.82.6 requires openai>=2.8.0.
# Install only after confirming a mem0ai release that supports openai v2:
#   uv pip install "mem0ai>=0.1.40,<0.2.0"
# EpisodicMemory in episodic.py degrades gracefully when mem0ai is absent.
episodic = ["mem0ai>=0.1.40,<0.2.0"]

[tool.uv.sources]
sovereign-edge-core = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/memory"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
